Trigger `on_container_exit` hook on docker `die` event.
