+++ virtool_workflow_runtime/cli.py
@@ -5,6 +5,8 @@ import logging
 # noinspection PyUnresolvedReferences
 import virtool_workflow_runtime._docker
 # noinspection PyUnresolvedReferences
+import virtool_workflow_runtime._docker_events
+# noinspection PyUnresolvedReferences
 import virtool_workflow_runtime._redis
 # noinspection PyUnresolvedReferences
 import virtool_workflow_runtime.cancellations
